refactor and review, icon button ok
