+++ app/src/main/java/com/piconemarc/personalaccountmanager/ui/baseComponent/Buttons.kt
@@ -1,126 +1,21 @@
 package com.piconemarc.personalaccountmanager.ui.baseComponent
 
-import androidx.compose.animation.ExperimentalAnimationApi
-import androidx.compose.animation.animateColorAsState
-import androidx.compose.animation.core.Spring
-import androidx.compose.animation.core.animateDpAsState
-import androidx.compose.animation.core.spring
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.CornerSize
-import androidx.compose.material.*
-import androidx.compose.runtime.*
+import androidx.compose.material.Button
+import androidx.compose.material.ButtonDefaults
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import com.piconemarc.personalaccountmanager.R
-import com.piconemarc.personalaccountmanager.ui.theme.*
+import com.piconemarc.personalaccountmanager.ui.theme.BigMarge
+import com.piconemarc.personalaccountmanager.ui.theme.ButtonShape
+import com.piconemarc.personalaccountmanager.ui.theme.PersonalAccountManagerTheme
+import com.piconemarc.personalaccountmanager.ui.theme.XlMarge
 
-@Composable
-fun PamIconButton(
-    iconButton: IconButtons,
-    onIconButtonClicked: () -> Unit
-) {
-    IconButton(onClick = onIconButtonClicked)
-    {
-        Surface(
-            modifier = Modifier
-                .background(MaterialTheme.colors.primary, CircleShape)
-                .padding(5.dp),
-            shape = CircleShape,
-            color = MaterialTheme.colors.primary,
-            border = BorderStroke(1.dp, MaterialTheme.colors.onPrimary),
-        )
-        {
-            Icon(
-                imageVector = ImageVector.vectorResource(iconButton.vectorIcon()),
-                contentDescription = stringResource(iconButton.contentDescription()),
-                modifier = Modifier
-                    .background(Color.Transparent, CircleShape)
-                    .padding(4.dp),
-                tint = MaterialTheme.colors.onPrimary
-            )
-        }
-    }
-}
-
-enum class IconButtons {
-    HOME {
-        override fun contentDescription(): Int {
-            return R.string.homeIconContentDescription
-        }
-
-        override fun vectorIcon(): Int {
-            return R.drawable.ic_outline_home_24
-        }
-    },
-    OPERATION {
-        override fun contentDescription(): Int {
-            return R.string.operationIconContentDescription
-        }
-
-        override fun vectorIcon(): Int {
-            return R.drawable.ic_outline_payments_24
-        }
-    },
-    PAYMENT {
-        override fun contentDescription(): Int {
-            return R.string.paymentIconContentDescription
-        }
-
-        override fun vectorIcon(): Int {
-            return R.drawable.ic_outline_ios_share_24
-        }
-    },
-    TRANSFER {
-        override fun contentDescription(): Int {
-            return R.string.transferIconContentDescription
-        }
-
-        override fun vectorIcon(): Int {
-            return R.drawable.ic_baseline_swap_horiz_24
-        }
-    },
-    CHART {
-        override fun contentDescription(): Int {
-            return R.string.chartIconContentDescription
-        }
-
-        override fun vectorIcon(): Int {
-            return R.drawable.ic_outline_bar_chart_24
-        }
-    },
-    ADD {
-        override fun contentDescription(): Int {
-            return R.string.addIconContentDescription
-        }
-
-        override fun vectorIcon(): Int {
-            return R.drawable.ic_baseline_add_24
-        }
-    };
-
-    abstract fun contentDescription(): Int
-    abstract fun vectorIcon(): Int
-}
-
-@Preview
-@Composable
-fun BaseIconPreview() {
-    PamIconButton(
-        onIconButtonClicked = {},
-        iconButton = IconButtons.CHART
-    )
-}
 
 @Composable
 fun BaseButton(
@@ -172,110 +67,4 @@ fun AcceptOrDismissPreview() {
     }
 }
 
-@Composable
-fun SwitchButton(
-    onButtonSelected: () -> Unit,
-    isSelected: Boolean,
-    title: String,
-    switchShape: Shape,
-    bottomPadding: Dp = RegularMarge
-) {
-    val buttonColor by animateColorAsState(targetValue = if (isSelected) MaterialTheme.colors.primary else MaterialTheme.colors.primaryVariant,animationSpec = if (!isSelected) tween(delayMillis = 120) else tween(delayMillis = 0) )
-    val textColor by animateColorAsState(targetValue = if (isSelected) MaterialTheme.colors.onPrimary else MaterialTheme.colors.onSecondary, animationSpec = if (!isSelected) tween(delayMillis = 120) else tween(delayMillis = 0))
-    Button(
-        modifier = Modifier
-            .wrapContentSize()
-            .background(
-                color = buttonColor,
-                shape = switchShape
-            )
-            .padding(end = RegularMarge),
-        onClick = onButtonSelected,
-        colors = ButtonDefaults.buttonColors(backgroundColor = Color.Transparent),
-        elevation = ButtonDefaults.elevation(0.dp)
 
-    ) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.h3,
-            modifier = Modifier.padding(
-                top = RegularMarge,
-                bottom = bottomPadding,
-                end = RegularMarge
-            ),
-            color = textColor
-        )
-    }
-}
-
-
-@OptIn(ExperimentalAnimationApi::class)
-@Composable
-fun PunctualOrRecurrentSwitchButton(
-    onEndDateSelected: (date: Pair<String, String>) -> Unit,
-    modifier : Modifier,
-    isRecurrent : (Boolean)->Unit
-) {
-    var selectedButton: Int by remember {
-        mutableStateOf(0)
-    }
-    var selectedDate: Pair<String, String> by remember {
-        mutableStateOf(Pair("", ""))
-    }
-    val recurrent = 1
-    val punctual = 0
-    isRecurrent ( selectedButton==1)
-
-    Column ( modifier = modifier.padding(bottom = RegularMarge)) {
-        Row(
-            horizontalArrangement = Arrangement.SpaceBetween,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(
-                    top = RegularMarge,
-                )
-        ) {
-            val shapeSize by animateDpAsState(targetValue = if (selectedButton == recurrent) BigMarge else RegularMarge, animationSpec = if (selectedButton != recurrent) tween(delayMillis = 120) else tween(delayMillis = 0) )
-            val shapeCorner by animateDpAsState(targetValue = if (selectedButton == recurrent) 0.dp else XlMarge, animationSpec = if (selectedButton != recurrent) tween(delayMillis = 120) else tween(delayMillis = 0))
-            SwitchButton(
-                onButtonSelected = { selectedButton = punctual },
-                isSelected = selectedButton == punctual,
-                title = stringResource(R.string.punctualSwitchButton),
-                switchShape = LeftSwitchShape
-            )
-            SwitchButton(
-                onButtonSelected = { selectedButton = recurrent },
-                isSelected = selectedButton == recurrent,
-                title = stringResource(R.string.recurrentSwitchButton),
-                switchShape =  RightSwitchShape.copy(
-                    bottomStart = CornerSize(
-                        shapeCorner
-                    )
-                ) ,
-                bottomPadding = shapeSize
-            )
-        }
-        //Recurrent options
-
-            RecurrentOptionPanel(
-                modifier = Modifier.height(expandCollapseEndDatePanel(isSelected = selectedButton == recurrent).value),
-                onMonthSelected = { month ->
-                    selectedDate = selectedDate.copy(first = month)
-                    onEndDateSelected(selectedDate)
-                },
-                onYearSelected = { year ->
-                    selectedDate = selectedDate.copy(second = year)
-                    onEndDateSelected(selectedDate)
-                }
-            )
-
-
-    }
-
-
-}
-@Composable
-fun expandCollapseEndDatePanel(isSelected : Boolean):State<Dp> = animateDpAsState(
-    targetValue = if (isSelected) 100.dp else 0.dp,
-    animationSpec = spring(dampingRatio = Spring.DampingRatioLowBouncy, stiffness = Spring.StiffnessLow)
-)
