+++ app/src/main/java/com/piconemarc/personalaccountmanager/ui/baseComponent/PopUp.kt
@@ -10,6 +10,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.piconemarc.personalaccountmanager.R
+import com.piconemarc.personalaccountmanager.ui.baseComponent.addOperationPopUp.AddOperationPopUp
+import com.piconemarc.personalaccountmanager.ui.baseComponent.addOperationPopUp.AddOperationPopUpState
 import com.piconemarc.personalaccountmanager.ui.theme.Positive
 import com.piconemarc.personalaccountmanager.ui.theme.deleteOperationTextModifier
