+++ app/src/main/java/com/piconemarc/personalaccountmanager/ui/baseComponent/PopUpAnimations.kt
@@ -10,16 +10,20 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.composed
 import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.piconemarc.personalaccountmanager.R
+import com.piconemarc.personalaccountmanager.ui.baseComponent.addOperationPopUp.AddOperationPopUpState
+import com.piconemarc.personalaccountmanager.ui.baseComponent.addOperationPopUp.TransitionData
+import com.piconemarc.personalaccountmanager.ui.baseComponent.addOperationPopUp.isRecurrent
 
 @SuppressLint("UnnecessaryComposedModifier")
-fun Modifier.selectorOffsetAnimation(selectedOperationOption: String): Modifier =
+fun Modifier.selectorOffsetAnimation(selectedOperationType: Int): Modifier =
     this.composed {
+
         val popUpLeftSidePanelIconSelectorTransition =
-            updateTransition(targetState = selectedOperationOption, label = "")
+            updateTransition(targetState = selectedOperationType, label = "")
+
         val selectorOffset by popUpLeftSidePanelIconSelectorTransition.animateOffset(
             label = "",
             transitionSpec = {
@@ -30,8 +34,8 @@ fun Modifier.selectorOffsetAnimation(selectedOperationOption: String): Modifier 
             }
         ) { selectedOperationOption_ ->
             when (selectedOperationOption_) {
-                stringResource(R.string.payment) -> Offset(0f, 58f)
-                stringResource(R.string.transfer) -> Offset(0f, 116f)
+                R.string.payment -> Offset(0f, 58f)
+                R.string.transfer -> Offset(0f, 116f)
                 else -> Offset(0f, 0f)
             }
         }
@@ -39,8 +43,8 @@ fun Modifier.selectorOffsetAnimation(selectedOperationOption: String): Modifier 
     }
 
 @Composable
-fun expandCollapsePaymentAnimation(popUpTitle: String): State<Dp> = animateDpAsState(
-    targetValue = if (popUpTitle != stringResource(id = R.string.operation)) {
+fun expandCollapsePaymentAnimation(popUpTitle: Int): State<Dp> = animateDpAsState(
+    targetValue = if (popUpTitle != R.string.operation) {
         if (!isRecurrent.value) 90.dp
         else 190.dp
     } else 0.dp,
@@ -51,8 +55,8 @@ fun expandCollapsePaymentAnimation(popUpTitle: String): State<Dp> = animateDpAsS
 )
 
 @Composable
-fun expandCollapseTransferAnimation(popUpTitle: String): State<Dp> = animateDpAsState(
-    targetValue = if (popUpTitle == stringResource(id = R.string.transfer)) 130.dp else 1.dp,
+fun expandCollapseTransferAnimation(popUpTitle: Int): State<Dp> = animateDpAsState(
+    targetValue = if (popUpTitle == R.string.transfer) 130.dp else 0.dp,
     animationSpec = spring(
         dampingRatio = Spring.DampingRatioLowBouncy,
         stiffness = Spring.StiffnessLow
